backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-stubs==1.13.0
django-cors-headers==3.13.0
PyMySQL==1.0.2
python-dotenv==0.21.0
